// generated/ts/sunshine.generated.ts
export const sunshineMeta: ThemeMeta = {
  name: 'sunshine',
  mode: 'light',
  family: 'duskmoon',
  pair: 'moonlight',
  description: 'Sunlit ivory with golden amber, muted lavender and sky blue',
};

export const sunshineColors: ThemeColors = {
  primary: '82.60893044% 0.1414340166 83.31858015',
  'primary-content': '28.28953512% 0.0448991160 80.63639668',
  'primary-container': '95.57475591% 0.0614713639 92.12872653',
  'on-primary-container': '38.72869883% 0.0701556434 81.75114485',
  secondary: '75.54657159% 0.0738680191 298.09280514',
  'secondary-content': '30.09347875% 0.0578718110 298.71400821',
  'secondary-container': '94.77129416% 0.0180917690 303.41223942',
  'on-secondary-container': '37.75435631% 0.0773343740 301.23088405',
  tertiary: '76.94385339% 0.0640284924 221.04480848',
  'tertiary-content': '30.82709477% 0.0391899830 225.86055898',
  'tertiary-container': '95.11671506% 0.0167909766 215.45868624',
  'on-tertiary-container': '37.64053450% 0.0461503944 221.94991006',
  accent: '90.26749798% 0.1105427306 93.87593258',
  'accent-content': '35.60101539% 0.0588941318 87.79310558',
  neutral: '34.78392925% 0.0173048639 78.03949200',
  'neutral-content': '98.35030566% 0.0170577608 84.58657624',
  'neutral-variant': '48.47267982% 0.0259043741 73.78871657',
  surface: '99.14473089% 0.0097633072 87.46952653',
  'surface-dim': '93.02387608% 0.0258398160 84.59321482',
  'surface-bright': '99.99999935% 0 0',
  'surface-container-lowest': '99.99999935% 0 0',
  'surface-container-low': '97.37543141% 0.0167153184 87.99753171',
  'surface-container': '95.32878442% 0.0224695782 87.15156842',
  'surface-container-high': '93.02387608% 0.0258398160 84.59321482',
  'surface-container-highest': '90.04844636% 0.0317371513 84.59011110',
  'surface-variant': '90.04844636% 0.0317371513 84.59011110',
  'on-surface': '28.54323950% 0.0202535180 76.57676498',
  'on-surface-variant': '48.47267982% 0.0259043741 73.78871657',
  'base-100': '99.14473089% 0.0097633072 87.46952653',
  'base-200': '97.37543141% 0.0167153184 87.99753171',
  'base-300': '93.02387608% 0.0258398160 84.59321482',
  'base-400': '55.01199157% 0.0351909086 74.88191474',
  'base-500': '44.93783706% 0.0295877832 74.85638734',
  'base-600': '35.00292970% 0.0249407737 76.85509951',
  'base-700': '24.86662243% 0.0194041154 73.70277146',
  'base-800': '18.07521115% 0.0118517040 73.14319887',
  'base-900': '11.73198826% 0.0090103504 73.47505765',
  'base-content': '28.54323950% 0.0202535180 76.57676498',
  outline: '57.87187705% 0.0396997009 75.20078373',
  'outline-variant': '85.83658493% 0.0325359084 82.75783875',
  'inverse-surface': '28.54323950% 0.0202535180 76.57676498',
  'inverse-on-surface': '98.35030566% 0.0170577608 84.58657624',
  'inverse-primary': '90.26749798% 0.1105427306 93.87593258',
  shadow: '0.00000000% 0 0',
  scrim: '0.00000000% 0 0 / 50%',
  info: '50.82237171% 0.0901951310 235.77699072',
  'info-content': '99.99999935% 0 0',
  'info-container': '95.09147873% 0.0151283794 228.96415080',
  'on-info-container': '39.58659448% 0.0609008792 232.98919080',
  success: '52.17594416% 0.0995342775 154.56738341',
  'success-content': '99.99999935% 0 0',
  'success-container': '94.93242285% 0.0208030926 140.89790892',
  'on-success-container': '39.51234051% 0.0596350615 156.43156459',
  warning: '55.13077016% 0.1354192110 53.04846323',
  'warning-content': '99.99999935% 0 0',
  'warning-container': '94.03448323% 0.0406430766 74.46372929',
  'on-warning-container': '40.76011391% 0.0954501128 54.71444706',
  error: '54.21677601% 0.1543098528 27.29045200',
  'error-content': '99.99999935% 0 0',
  'error-container': '93.88397486% 0.0266133029 34.97748526',
  'on-error-container': '38.73080591% 0.0993593486 27.58854513',
};

export const sunshineShape: ThemeShape = {
  'radius-selector': '0.25rem',
  'radius-field': '0.625rem',
  'radius-box': '2rem',
  'size-selector': '0.1875rem',
  'size-field': '0.1875rem',
  border: '0.5px',
  depth: 1,
  noise: 1,
};